requirements.txt
faiss_cpu
gensim
importlib_resources
nltk
numba
numpy
pandas
scikit_learn
scipy
tqdm
packaging